test(autosave-association): move Rails-less tests to the trails sibling file
